custom devcon location

+++ src/modules/settings.py
@@ -1,4 +1,5 @@
 import json
+from os import path
 from PyQt6.QtWidgets import QDialog
 from PyQt6.QtGui import QIcon
 from assets.settings_ui import Ui_settingsDialog
@@ -6,7 +7,7 @@ from src.helpers.helper import (
   verify_settings,
   get_default_settings
 )
-
+from src.helpers.errors import Errors
 
 class Settings(QDialog, Ui_settingsDialog):
   # labels: hidLabel, devconLabel
@@ -26,14 +27,30 @@ class Settings(QDialog, Ui_settingsDialog):
     try:
       settingsPath = verify_settings(self.pwd)
       hidDeviceId = str(self.hidText.text()).strip()
+      devconPath = str(self.devconText.text()).strip()
+
+      if not (path.isfile(devconPath) or devconPath == ''):
+        Errors.raise_error(
+          self,
+          'SETTINGS_TITLE',
+          'MISSING_DEVCON_ERROR',
+          'WARNING'
+        )
+        return
 
       structure = get_default_settings()
       structure['hid_device_id'] = hidDeviceId
+      structure['devcon_path'] = devconPath
       with open(settingsPath, 'w', encoding="utf-8") as file:
         json.dump(structure, file)
       self.close_settings_ui()
     except Exception as e:
-      print(e)
+      Errors.raise_error(
+        self,
+        'SETTINGS_TITLE',
+        'SAVE_SETTINGS_ERROR',
+        'WARNING'
+      )
 
   def load_settings(self):
     try:
@@ -43,8 +60,14 @@ class Settings(QDialog, Ui_settingsDialog):
         settings = json.load(file)
 
       self.hidText.setText(settings['hid_device_id'])
+      self.devconText.setText(settings['devcon_path'])
     except Exception as e:
-      print(e)
+      Errors.raise_error(
+        self,
+        'SETTINGS_TITLE',
+        'LOAD_SETTINGS_ERROR',
+        'WARNING'
+      )
 
   def load_settings_ui(self):
     self.load_settings()
